Refactor into widgets module
