+++ widgets/__init__.py
@@ -1,0 +1,7 @@
+from widgets.widget_base import Widget
+from widgets.dataset_description import DatasetDescription
+from widgets.general_stats import GeneralStats
+from widgets.label_distribution import LabelDistribution
+from widgets.npmi import Npmi
+from widgets.text_lengths import TextLengths
+from widgets.zipf import Zipf
